Moved Searchbar into own component, refactored Navbar
