+++ src/Components/MapComponents/Navbar.js
@@ -5,41 +5,13 @@ import { connect } from "react-redux";
 import "./Navbar.scss";
 import { updateChapters } from "../../Actions/SearchBarAction";
 
-// search bar component
-// import SearchBar from "../MapComponents/SearchBar.js";
+import SearchBar from "../MapComponents/SearchBar.js";
 import axios from "axios";
 import { useOktaAuth } from '@okta/okta-react';
 
 import profilephoto from '../../Assets/Imgs/USER-PROF.png';
-import { Media } from "reactstrap";
 
-const Navbar = (props) => {
-  const [search, updateSearch] = useState("");
-  const [chapters, updateChapters] = useState([]);
-
-  useEffect(() => {
-    axios
-      .get("https://miracle-messages-dev.herokuapp.com/api/chapter", chapters)
-      .then((res) => {
-        updateChapters(res.data);
-      })
-      .catch((err) => {
-        console.log("search", err);
-      });
-  }, []);
-
-  const handleChange = (e) => {
-    updateSearch(e.target.value);
-  };
-
-  useEffect(() => {
-    props.updateChapters(
-      chapters.filter((chapter) =>
-        chapter.city.toLowerCase().includes(search.toLowerCase())
-      )
-    );
-  }, [search]);
-
+const Navbar = () => {
   return (
     <div className="navbar-map">
       <Link to="/">
@@ -52,23 +24,7 @@ const Navbar = (props) => {
           src={logo} alt="logo" />
       </Link>
       <nav>
-        <div className="search-box">
-          {/* move search bar into separate component */}
-          <div className="search-bar">
-            <form>
-              <i className="fas fa-search" />
-              &nbsp;&nbsp;
-              <input
-                type="text"
-                placeholder="Search Chapters"
-                value={search}
-                onChange={handleChange}
-                //style=()
-                className="input"
-              />
-            </form>
-          </div>
-        </div>
+        <SearchBar />
         <div style={{ margin: "15px" }}> </div>
         
         <Link to="/">Dashboard</Link>
@@ -105,11 +61,6 @@ const Navbar = (props) => {
       </nav>
     </div>
   );
-};
+ };
 
-const mapStateToProps = (state) => {
-  return {
-    chapterData: state.mapReducer.chapterData,
-  };
-};
-export default connect(mapStateToProps, { updateChapters })(Navbar);
+export default Navbar;
